project/requirements.txt
streamlit
requests
python-dotenv
pillow
flask
azure-cognitiveservices-speech
reportlab
PyPDF2
moviepy
matplotlib
seaborn
openai
fpdf
pydub
yt-dlp
wave
ffmpeg